Increase read timeout to 10 seconds
To keep the connection open, for instance, when doing a 'pacman -Syu',
we wait for the client to confirm the update. This will take a few
seconds. If the interval is too short, the client won't see any errors,
but performance is slightly worse since a new connection needs to be
established.