Remove useless resume_from/cached_size variable
We're already using the progress_indicator() function to check if the
file is already partially available. The cached_size variable was
removed because this variable does not have the correct size if multiple
download attempts are required: For example, suppose that the download
starts from mirror A, and then 4096 bytes are fetched from this mirror.
Then, the mirror turns out to be too slow, so we switch to another
mirror. In this case, the variable cached_size still has the old size
(i.e., 0) instead of 4096. The progress_indicator() function, on the
other hand, has the correct size.

This was not a bug, because we have previously used the value from the
progress_indicator() function in a situation like the one described
above. However, it is just needlessly complicated, so we remove the
cached_size variable.